+++ src/frontend/web_application/server/auth/middlewares/decode-cookie.ts
@@ -2,6 +2,11 @@ import { decode } from '../lib/seal';
 import { COOKIE_NAME } from '../lib/cookie';
 import { getConfig } from '../../config';
 
+class ResponseError extends Error {
+  status: number;
+  err: any;
+}
+
 export const decodeCookieMiddleware = (req, res, next) => {
   const {
     seal: { secret },
@@ -14,22 +19,23 @@ export const decodeCookieMiddleware = (req, res, next) => {
     return;
   }
 
-  decode(cookie, secret, (err, obj) => {
-    if (err || !obj) {
-      const error = new Error('Unexpected Server Error on cookie decoding');
+  decode(cookie, secret).then(
+    (obj) => {
+      req.user = obj;
+
+      // TODO: refresh token
+      // new Date(Date.UTC(req.user.tokens.expires_at)).getTime() < new Date().getTime() + (1000 * 60 * 10)
+
+      next();
+    },
+    (err) => {
+      const error = new ResponseError(
+        'Unexpected Server Error on cookie decoding'
+      );
       error.status = 500;
       error.err = err;
 
       next(error);
-
-      return;
     }
-
-    req.user = obj;
-
-    // TODO: refresh token
-    // new Date(Date.UTC(req.user.tokens.expires_at)).getTime() < new Date().getTime() + (1000 * 60 * 10)
-
-    next();
-  });
+  );
 };