+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to
 [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Fixed
+
+- Update frontend server dependency to correctly handle cookie decoding.
+
 ## [0.24.4] 2021-07-25
 
 ### Fixed